requirements.txt
# Core dependencies
python-telegram-bot==21.10
aiohttp==3.11.11
python-dotenv==1.0.1

# LLM Integration (OpenAI compatible con Ollama)
langchain==1.0.1
langchain-openai==1.0.0
langchain-core==1.0.0

# Docker & Monitoring
docker==7.1.0
psutil==6.1.1

# Utils
pyyaml==6.0.2
pydantic==2.10.5
